LPD-101603 Extract the shared client extension page objects
ClientExtensionsPage, EditClientExtensionsPage, and ViewClientExtensionPage
import one another and, together with the clientExtensionsPageTest fixture that
wraps the first of them, are used by nearly every spec in the component. They
sat under main, so the upcoming per-client-extension-type projects would have
had to reach into a sibling project to use them.

They now live at the component level, which the Playwright suite already uses
for shared code in object-web and commerce. Three specs outside the component
were already importing them across the tree, so they gain a supported location
rather than a path into another project's internals.
