+++ modules/test/playwright/tests/client-extension-web/cluster/cluster.spec.ts
@@ -9,7 +9,7 @@ import {loginTest} from '../../../fixtures/loginTest';
 import {liferayConfig} from '../../../liferay.config';
 import getRandomString from '../../../utils/getRandomString';
 import {editJSImportMapsPageTest} from '../main/fixtures/editJSImportMapsExtensionPageTest';
-import {WaitAction} from '../main/pages/EditClientExtensionsPage';
+import {WaitAction} from '../pages/EditClientExtensionsPage';
 
 export const test = mergeTests(editJSImportMapsPageTest, loginTest());
