+++ modules/test/playwright/tests/client-extension-web/fixtures/clientExtensionsPageTest.ts
