+++ modules/test/playwright/tests/client-extension-web/main/jsImportMapsEntry.spec.ts
@@ -7,10 +7,10 @@ import {expect, mergeTests} from '@playwright/test';
 
 import {loginTest} from '../../../fixtures/loginTest';
 import getRandomString from '../../../utils/getRandomString';
-import {clientExtensionsPageTest} from './fixtures/clientExtensionsPageTest';
+import {clientExtensionsPageTest} from '../fixtures/clientExtensionsPageTest';
+import {WaitAction} from '../pages/EditClientExtensionsPage';
+import {ViewClientExtensionPage} from '../pages/ViewClientExtensionPage';
 import {editJSImportMapsPageTest} from './fixtures/editJSImportMapsExtensionPageTest';
-import {WaitAction} from './pages/EditClientExtensionsPage';
-import {ViewClientExtensionPage} from './pages/ViewClientExtensionPage';
 
 const test = mergeTests(
 	clientExtensionsPageTest,
