+++ modules/test/playwright/tests/client-extension-web/main/pages/EditJSClientExtensionsPage.ts
@@ -6,7 +6,7 @@
 import {Locator, Page} from '@playwright/test';
 
 import {clickAndExpectToBeHidden} from '../../../../utils/clickAndExpectToBeHidden';
-import {EditClientExtensionsPage} from './EditClientExtensionsPage';
+import {EditClientExtensionsPage} from '../../pages/EditClientExtensionsPage';
 
 export class EditJSClientExtensionsPage extends EditClientExtensionsPage {
 	readonly javaScriptURLInput: Locator;
