+++ modules/test/playwright/tests/client-extension-web/main/pages/EditJSImportMapsPage.ts
@@ -5,7 +5,7 @@
 
 import {Locator, Page} from '@playwright/test';
 
-import {EditClientExtensionsPage} from './EditClientExtensionsPage';
+import {EditClientExtensionsPage} from '../../pages/EditClientExtensionsPage';
 
 export class EditJSImportMapsPage extends EditClientExtensionsPage {
 	readonly bareSpecifierInput: Locator;
