+++ modules/test/playwright/tests/client-extension-web/main/pages/EditThemeSVGSpritemapPage.ts
@@ -5,7 +5,7 @@
 
 import {Locator, Page} from '@playwright/test';
 
-import {EditClientExtensionsPage} from './EditClientExtensionsPage';
+import {EditClientExtensionsPage} from '../../pages/EditClientExtensionsPage';
 
 export class EditThemeSVGSpritemapPage extends EditClientExtensionsPage {
 	readonly urlInput: Locator;
