+++ modules/test/playwright/tests/portal-workflow-kaleo-designer-web/main/workflowDefinitions.spec.ts
@@ -12,7 +12,7 @@ import {workflowPagesTest} from '../../../fixtures/workflowPagesTest';
 import {clickAndExpectToBeHidden} from '../../../utils/clickAndExpectToBeHidden';
 import getRandomString from '../../../utils/getRandomString';
 import getBasicWebContentStructureId from '../../../utils/structured-content/getBasicWebContentStructureId';
-import {clientExtensionsPageTest} from '../../client-extension-web/main/fixtures/clientExtensionsPageTest';
+import {clientExtensionsPageTest} from '../../client-extension-web/fixtures/clientExtensionsPageTest';
 import {journalPagesTest} from '../../journal-web/main/fixtures/journalPagesTest';
 import {getWorkflowDefinition} from './utils/getWorkflowDefinition';
