+++ modules/test/playwright/tests/users-admin-web/main/customElement.spec.ts
@@ -10,7 +10,7 @@ import {loginTest} from '../../../fixtures/loginTest';
 import {liferayConfig} from '../../../liferay.config';
 import {getRandomInt} from '../../../utils/getRandomInt';
 import {editCustomElementPageTest} from '../../client-extension-web/main/fixtures/editCustomElementPageTest';
-import {WaitAction} from '../../client-extension-web/main/pages/EditClientExtensionsPage';
+import {WaitAction} from '../../client-extension-web/pages/EditClientExtensionsPage';
 import {componentsPageTest} from '../../configuration-admin-web/main/fixtures/ComponentsPageTest';
 
 export const test = mergeTests(
